Bug 13539: Remove unused table patroncards
It appears that the table patroncards hasn't been used
since the database updates for the rewritten label modules
('3.01.00.107').

It's time to remove them from the database.

As the patron card batches were needed for printing and
the data has not been accessible in a long time, it seems safe
to delete it.

Signed-off-by: Brendan Gallagher <brendan@bywatersolutions.com>

Signed-off-by: Jonathan Druart <jonathan.druart@biblibre.com>
Signed-off-by: Tomas Cohen Arazi <tomascohen@gmail.com>