Bug 13486: Apache version empty in about page if Apache 2.4
This patch makes about.pl query the running Apache version using
the apache2ctl command. I tested it on Apache 2.2 and is backwards
compatible (it is present in both).

To test:
( On an Apache 2.4 setup )
- Go to the about page
=> FAIL: "Apache version:" is empty
- Apply the patch
- Reload the page
=> SUCCESS: "Apache version:" shows the correct version.
- Sign off :-D

If you have an Apache 2.2 setup to try it, verify that there are no regressions
(trivial because you can run apache2ctl manually and verify the first result is the expecte).

Regards

PS. I even changed the order in which Apache version is tested because
most people is using Debian/Ubuntu and it was the last option.

Signed-off-by: Bernardo Gonzalez Kriegel <bgkriegel@gmail.com>
Tested with apache 2.2/2.4, no errors

Signed-off-by: Jonathan Druart <jonathan.druart@koha-community.org>
Only tested with 2.4
Signed-off-by: Tomas Cohen Arazi <tomascohen@gmail.com>