Bug 13657 - Don't show extra comma after guarantor name on patron details
This patch suppresses the first name part (including comma) if first name is empty. This way, first name is still displays with natural persons, but not with organizations.

To test:
Apply this patch (without the first patch).
Have one patron with a natural person as guarantor and one patron with an organization as guarantor. Verify that both display properly on the patron details page.

Signed-off-by: Jesse Maseto <Jesse@bywatersolutions.com>
Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>
Signed-off-by: Tomas Cohen Arazi <tomascohen@gmail.com>