Bug 11900 - OPAC cart can confuse patrons
This patch adds a title tag to the "Cart" link on Opac main page (and to the "Lists" link as well).

Wording should be verified by a native speaker  :-)

To test:
Apply patch
Go to OPAC main page, verify that a tooltip appears on the "Cart" and "Lists" links at the top of the page.

Signed-off-by: Nick <nick@quecheelibrary.org>
Signed-off-by: Jesse Maseto <jesse@bywatersolutions.com>
Signed-off-by: Katrin Fischer <Katrin.Fischer.83@web.de>
Works as described, no problems found.

Signed-off-by: Tomas Cohen Arazi <tomascohen@gmail.com>