Bug 13866: Columns configuration - Patrons search
This patch adds the ability to hide columns on the patrons search page
(acqui/lateorders.pl).

Test plan:
Go on the patrons search page and confirm you are able to hide columns.
You can also play with the columns configuration admin page and confirm
that the behaviors are consistent.

Signed-off-by: Josef Moravec <josef.moravec@gmail.com>
Works as expected

Signed-off-by: Kyle M Hall <kyle@bywatersolutions.com>
Signed-off-by: Tomas Cohen Arazi <tomascohen@gmail.com>
Small rebase was needed.