Bug 13923: The prefetch used on existing constraint
The belongs_to constraint has been removed by bug 13713.
Not sure if it was a good idea, but atm the easy and quick way is to
remove the prefetch.

Test plan:
Go on the circulation page, holds tab
Confirm that there is no regression
(Same on the patron details page > Holds tab)

Signed-off-by: Kyle M Hall <kyle@bywatersolutions.com>

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>
Signed-off-by: Tomas Cohen Arazi <tomascohen@gmail.com>