Bug 9333: Remove useless JS code - confirm close basket
On closing a basket, the librarian is redirect to a new page for the
confirmation, the JavaScript confirmation is useless.

This code is never triggered and can be removed.

Test plan:
Without this patch, confirm you don't get a JavaScript confirmation
popup on closing a basket
With this patch, confirm you don't get any JavaScript error on
closing/reopening a basket.

NOTE: acqui/basket.tt is only used by acqui/basket.pl
      git grep "confirm_close" | grep -v "\.po:"
      doesn't show anything else. Reopening only works on
      basketgroups, not baskets. Unable to trigger any errors
      as expected and desired. :)

Signed-off-by: Mark Tompsett <mtompset@hotmail.com>

Signed-off-by: Kyle M Hall <kyle@bywatersolutions.com>
Signed-off-by: Tomas Cohen Arazi <tomascohen@gmail.com>