Bug 10947: Fix editing the default authority type
The default authority type cannot be defined.
The authtypecode is en empty string and the tests should be done on
"defined", not "exist".

Test plan:
Edit the default authority type, save it.

Note: There is no way to create an authority with an empty string, with
you deleted you won't be able to recreate it.

Followed test plan, works as expected.
Signed-off-by: Marc Véron <veron@veron.ch>

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>
Signed-off-by: Tomas Cohen Arazi <tomascohen@gmail.com>