Bug 14051: Tiny adjustment of ReturnPathDefault pref description
Adding that Return-Path is bounce address. Note that if it is empty, the From address is used. This could be the admin address, but Koha uses other from addresses too.

Test plan:
Read the description on the preferences page.
Check some Koha mails with/without this pref. Note that this patch does not
change behavior in any way.

Signed-off-by: Marc Veron <veron@veron.ch>

Signed-off-by: Kyle M Hall <kyle@bywatersolutions.com>
Signed-off-by: Tomas Cohen Arazi <tomascohen@gmail.com>