Bug 11890: Click events on titles does not work as expected
The browse results feature (bug 10404) modified the click events on
titles on the search result page (intranet).

It introduced bad behaviors:
1/ ctrl+click on a title does not open a new tab
2/ middle click on a title does open a new tab but the browser is not
displayed.

Note that this patch is not perfect, it fixes the 2 bad behaviors but
the ctrl+click gives the focus to the new tab (could go against the
preferences of the users).

Test plan:
1/ On the staff interface, launch a search (catalogue/search.pl?q=XXX)
2/ When middle-clicking on a title, a new tab should open. On this tab
the browse result feature should be displayed
3/ Same for cltr+click

Signed-off-by: Nick Clemens <nick@quecheelibrary.org>

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>
Signed-off-by: Tomas Cohen Arazi <tomascohen@gmail.com>